simplify integration test
