+++ scripts/client_test.sh
@@ -1,4 +1,18 @@
 #!/usr/bin/env bash
+# Copyright 2020 The SQLFlow Authors. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
 
 JOB_TYPE=$1
 PS_NUM=$2
@@ -6,12 +20,12 @@ WORKER_NUM=$3
 
 # Generate checkpoint for mnist to test evaluation and prediction
 MNIST_CKPT_DIR=model_zoo/test_data/mnist_ckpt/
-python -m scripts.gen_mnist_checkpoint --checkpoint_dir=${MNIST_CKPT_DIR}
+python scripts/gen_mnist_checkpoint.py --checkpoint_dir=${MNIST_CKPT_DIR}
 
 
 if [[ "$JOB_TYPE" == "train" ]]; then
     elasticdl train \
-      --image_base=elasticdl:ci \
+      --image_name=elasticdl:ci \
       --model_zoo=model_zoo \
       --model_def=deepfm_functional_api.deepfm_functional_api.custom_model \
       --training_data=/data/frappe/train \
@@ -25,8 +39,8 @@ if [[ "$JOB_TYPE" == "train" ]]; then
       --ps_resource_limit="cpu=1,memory=2048Mi" \
       --minibatch_size=64 \
       --num_minibatches_per_task=2 \
-      --num_workers=$WORKER_NUM \
-      --num_ps_pods=$PS_NUM \
+      --num_workers="$WORKER_NUM" \
+      --num_ps_pods="$PS_NUM" \
       --checkpoint_steps=500 \
       --evaluation_steps=500 \
       --tensorboard_log_dir=/tmp/tensorboard-log \
@@ -39,7 +53,7 @@ if [[ "$JOB_TYPE" == "train" ]]; then
       --volume="host_path=${PWD},mount_path=/saved_model"
 elif [[ "$JOB_TYPE" == "evaluate" ]]; then
     elasticdl evaluate \
-      --image_base=elasticdl:ci \
+      --image_name=elasticdl:ci \
       --model_zoo=model_zoo \
       --model_def=mnist_functional_api.mnist_functional_api.custom_model \
       --checkpoint_dir_for_init=/${MNIST_CKPT_DIR}/version-100  \
@@ -53,8 +67,8 @@ elif [[ "$JOB_TYPE" == "evaluate" ]]; then
       --ps_resource_limit="cpu=1,memory=2048Mi" \
       --minibatch_size=64 \
       --num_minibatches_per_task=2 \
-      --num_workers=$WORKER_NUM \
-      --num_ps_pods=$PS_NUM \
+      --num_workers="$WORKER_NUM" \
+      --num_ps_pods="$PS_NUM" \
       --evaluation_steps=15 \
       --tensorboard_log_dir=/tmp/tensorboard-log \
       --job_name=test-evaluate \
@@ -62,7 +76,7 @@ elif [[ "$JOB_TYPE" == "evaluate" ]]; then
       --image_pull_policy=Never
 elif [[ "$JOB_TYPE" == "predict" ]]; then
     elasticdl predict \
-      --image_base=elasticdl:ci \
+      --image_name=elasticdl:ci \
       --model_zoo=model_zoo \
       --model_def=mnist_functional_api.mnist_functional_api.custom_model \
       --checkpoint_dir_for_init=/${MNIST_CKPT_DIR}/version-100 \
@@ -75,8 +89,8 @@ elif [[ "$JOB_TYPE" == "predict" ]]; then
       --ps_resource_limit="cpu=1,memory=2048Mi" \
       --minibatch_size=64 \
       --num_minibatches_per_task=2 \
-      --num_workers=$WORKER_NUM \
-      --num_ps_pods=$PS_NUM \
+      --num_workers="$WORKER_NUM" \
+      --num_ps_pods="$PS_NUM" \
       --job_name=test-predict \
       --log_level=INFO \
       --image_pull_policy=Never
@@ -95,10 +109,10 @@ elif [[ "$JOB_TYPE" == "local" ]]; then
       --distribution_strategy=Local 
 elif [[ "$JOB_TYPE" == "odps" ]]; then
     elasticdl train \
-      --image_base=elasticdl:ci \
+      --image_name=elasticdl:ci \
       --model_zoo=model_zoo \
       --model_def=odps_iris_dnn_model.odps_iris_dnn_model.custom_model \
-      --training_data=$MAXCOMPUTE_TABLE \
+      --training_data="$MAXCOMPUTE_TABLE" \
       --data_reader_params='columns=["sepal_length", "sepal_width", "petal_length", "petal_width", "class"]; label_col="class"' \
       --envs="MAXCOMPUTE_PROJECT=$MAXCOMPUTE_PROJECT,MAXCOMPUTE_AK=$MAXCOMPUTE_AK,MAXCOMPUTE_SK=$MAXCOMPUTE_SK,MAXCOMPUTE_ENDPOINT=" \
       --num_epochs=2 \
@@ -110,8 +124,8 @@ elif [[ "$JOB_TYPE" == "odps" ]]; then
       --ps_resource_limit="cpu=1,memory=2048Mi" \
       --minibatch_size=64 \
       --num_minibatches_per_task=2 \
-      --num_workers=$WORKER_NUM \
-      --num_ps_pods=$PS_NUM \
+      --num_workers="$WORKER_NUM" \
+      --num_ps_pods="$PS_NUM" \
       --checkpoint_steps=10 \
       --grads_to_wait=2 \
       --job_name=test-odps \
