+++ scripts/gen_mnist_checkpoint.py
@@ -1,8 +1,23 @@
+# Copyright 2020 The SQLFlow Authors. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import argparse
 
 import tensorflow as tf
 
-from elasticdl.python.tests.test_utils import save_checkpoint_without_embedding
+from elasticdl.python.common.save_utils import (
+    save_checkpoint_without_embedding,
+)
 
 
 def mnist_custom_model():
